First attempt at #10
Turns out the obj format is *lot* more complex than can be handled by
simple line splitting and regexes (it's a line-based format with
backslash continuations, arbitrary comments, and a certain amount of
state based parsing like the negative index handling). This is an
attempt at using pyparsing to handle the format; this works but it's
*really* slow ... probably too slow to use on a Pi (it's taking 6
seconds to parse a 300kb model on my Core i7; I dread to think what
that translates to on a Pi 1).

Then again, pyparsing is pretty slow generally - might be worth trying
to hand-roll a lexer & parser for this?