Next stab at #10
After digging out the most complete docs on the .obj spec (appendix from
the Wavefront manual) it seems this format *can* be parsed with simple
line splitting provided one assumes (like the Blender variant of the
parser, and the vim syntax file) that commands all start in column 1.
This isn't actually in the spec, but oh well - enough existing parsers
agree with it to make it a de-facto standard.

The parser is now largely complete along with associated machinery, and
it's about as fast as we can reasonably hope to make it on a Pi (at
least it's much faster than the 2 minutes that the pyparsing variant
took...). Now to finish the render method, and we can consider this
done.