Fix #13
Implemented hollow sphere generator. Cube et al can wait until 1.0 and
the refactoring to implement memoized class-based shape generators. Also
fixed up the block setter so that it'll accept mis-matched ranges (which
in turn allows infinite cycles to be used when unknown length generators
are passed to it, as with the vector shape generators), and make circle
a bit nicer by warning about non-vector radii.